move interface into interface namespace
